# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  displayName: string | null;
  role: UserRole;
}

export interface AdminStats {
  userCount: number;
  adminCount: number;
  reviewCount: number;
  averageRating: number;
  assignmentCount: number;
  sessionCount: number;
  signupsLast7Days: number;
}

export type AdminUserRole = typeof AdminUserRole[keyof typeof AdminUserRole];


export const AdminUserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface AdminUser {
  id: number;
  email: string;
  displayName: string | null;
  role: AdminUserRole;
  createdAt: string;
  reviewCount: number;
  assignmentCount: number;
}

export interface AdminUserList {
  users: AdminUser[];
}

export type SetRoleInputRole = typeof SetRoleInputRole[keyof typeof SetRoleInputRole];


export const SetRoleInputRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface SetRoleInput {
  role: SetRoleInputRole;
}

export interface AuthResult {
  token: string;
  user: User;
}

export interface RegisterInput {
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
  displayName?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AppleAuthInput {
  /** @minLength 1 */
  identityToken: string;
}

export interface GoogleAuthInput {
  /** @minLength 1 */
  idToken: string;
}

export interface Review {
  id: number;
  rating: number;
  body: string | null;
  authorName: string | null;
  createdAt: string;
  mine: boolean;
}

export interface ReviewSummary {
  average: number;
  count: number;
  reviews: Review[];
  myReview: Review | null;
}

export interface ReviewInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @maxLength 2000 */
  body?: string;
}

export type AssignmentStatus = typeof AssignmentStatus[keyof typeof AssignmentStatus];


export const AssignmentStatus = {
  on_track: 'on_track',
  at_risk: 'at_risk',
  complete: 'complete',
} as const;

export interface Assignment {
  id: number;
  title: string;
  subject: string;
  dueDate: string;
  dueLabel: string;
  totalMinutes: number;
  completedMinutes: number;
  status: AssignmentStatus;
  accent: string;
}

export type StudyTaskStatus = typeof StudyTaskStatus[keyof typeof StudyTaskStatus];


export const StudyTaskStatus = {
  todo: 'todo',
  done: 'done',
} as const;

export interface StudyTask {
  id: number;
  assignmentId: number;
  title: string;
  durationMinutes: number;
  status: StudyTaskStatus;
  sortOrder: number;
}

export type StudySessionStatus = typeof StudySessionStatus[keyof typeof StudySessionStatus];


export const StudySessionStatus = {
  scheduled: 'scheduled',
  complete: 'complete',
  skipped: 'skipped',
} as const;

export interface StudySession {
  id: number;
  taskId: number;
  assignmentId: number;
  title: string;
  subject: string;
  date: string;
  startTime: string;
  endTime: string;
  durationMinutes: number;
  status: StudySessionStatus;
  isToday: boolean;
}

export interface TodayFocus {
  sessionId: number;
  taskId: number;
  title: string;
  subject: string;
  startTime: string;
  durationMinutes: number;
  context: string;
}

export interface WorkloadDay {
  date: string;
  dayLabel: string;
  minutes: number;
  sessionCount: number;
  isToday: boolean;
}

export interface PlannerDashboard {
  greeting: string;
  dateLabel: string;
  todayFocus: TodayFocus | null;
  todaySessions: StudySession[];
  upcomingAssignments: Assignment[];
  workload: WorkloadDay[];
  totalMinutesThisWeek: number;
  completedSessions: number;
  streakDays: number;
}

export interface StudyPlanInput {
  /** @minLength 3 */
  note: string;
  /**
     * @minimum 15
     * @maximum 480
     */
  availableMinutesPerDay?: number;
}

export interface StudyPlan {
  assignments: Assignment[];
  tasks: StudyTask[];
  sessions: StudySession[];
  summary: string;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  done: 'done',
} as const;

export interface TaskUpdate {
  status: TaskUpdateStatus;
}

export interface SessionRescheduleInput {
  date: string;
}