# lib/api-zod/src/generated/types/appleAuthInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppleAuthInput {
  /** @minLength 1 */
  identityToken: string;
}